test(backend): consolidate metricsCalculator + segmentationService test clusters
Aggressively de-duplicates two heavily-accreted backend test clusters into
coherent per-concern files with shared fixtures, no coverage loss.

- metricsCalculator: 6 files -> 3 (merged main/basicMetrics/gaps/gaps5 into
  one; kept sperm + microcapsule as distinct export-path files). 120 -> 103
  passing: removed exact-duplicate scale tests, shallow Array.isArray checks,
  a fake test that stubbed the SUT, one stale .skip; folded assertion-equal
  cases. All DI/panel/referenceMode coverage retained.
- segmentationService: 14 files -> 4 (core/batch/stats/SSOT/channel-path in
  the main file; kept concurrent; merged trackOps + trackOpsService +
  crossFrame; kept sperm). 174 -> 138 passing: removed cross-file duplicates
  across gaps/gaps3/gaps4/gaps5 and a 7-test all-skip integration file that
  targeted a non-existent API. Every distinct behaviour/branch retained.

Full backend suite: 197 files, 3599 passing, 4 skipped (was 210/3652/12);
the -53 passing / -8 skipped are exactly the removed duplicates/stale skips.
Net -3272 lines.
