--- backend/src/services/__tests__/segmentationService.gaps.test.ts
@@ -1,881 +1,0 @@
-/**
- * segmentationService.gaps.test.ts
- *
- * Covers the uncovered paths from segmentationService.ts:
- *  - saveSegmentationResultsInternal: polygon validation filtering, polyline
- *    merging (sperm model), invalid image_size, parent_id round-trip,
- *    trackId/name preservation via the public saveSegmentationResults wrapper.
- *  - getSegmentationResults: trackId/name preserved; _embedding stripped;
- *    parent_id→parentIds; updatedAt surfaced.
- *  - updateSegmentationResults: new-segmentation create path; existing update
- *    path; parentIds→parent_id conversion; trackId/name round-trip.
- *  - deleteSegmentationResults: success; image-not-found throw.
- *  - getProjectSegmentationStats: project not found; model usage tallying.
- */
-import { describe, it, expect, beforeEach, vi } from 'vitest';
-import { PrismaClient } from '@prisma/client';
-import { SegmentationService } from '../segmentationService';
-import { ImageService } from '../imageService';
-
-vi.mock('../../utils/logger', () => ({
-  logger: {
-    debug: vi.fn(),
-    info: vi.fn(),
-    warn: vi.fn(),
-    error: vi.fn(),
-  },
-}));
-
-vi.mock('../../utils/config', () => ({
-  config: {
-    SEGMENTATION_SERVICE_URL: 'http://localhost:8000',
-    STORAGE_TYPE: 'local',
-    STORAGE_LOCAL_PATH: '/tmp/test-storage',
-    UPLOAD_DIR: '/tmp/uploads',
-    NODE_ENV: 'test',
-  },
-}));
-
-vi.mock('../../storage');
-vi.mock('../segmentationThumbnailService');
-vi.mock('../thumbnailManager', () => ({
-  ThumbnailManager: function MockThumbnailManager(this: any) {
-    this.generateAllThumbnails = vi.fn().mockResolvedValue(undefined);
-  },
-}));
-vi.mock('../imageService');
-
-// ─── helpers ─────────────────────────────────────────────────────────────────
-
-const makePolygon = (
-  overrides: Record<string, unknown> = {}
-): Record<string, unknown> => ({
-  id: 'poly-1',
-  points: [
-    { x: 0, y: 0 },
-    { x: 10, y: 0 },
-    { x: 10, y: 10 },
-  ],
-  type: 'external',
-  area: 50,
-  confidence: 0.9,
-  ...overrides,
-});
-
-const makeSegResult = (
-  polygons: unknown[],
-  extra: Record<string, unknown> = {}
-) => ({
-  success: true,
-  polygons: polygons as any,
-  model_used: 'hrnet',
-  threshold_used: 0.5,
-  processing_time: 1.5,
-  image_size: { width: 800, height: 600 },
-  ...extra,
-});
-
-// getProjectSegmentationStats now uses server-side aggregation
-// (segmentation.aggregate + segmentation.groupBy + a $queryRaw polygon count)
-// instead of loading all rows into Node. Derive those mock returns from a flat
-// array of segmentation rows so the per-test data stays readable.
-const mockSegStatsFromRows = (
-  prisma: any,
-  rows: Array<{
-    polygons?: string;
-    confidence?: number | null;
-    model?: string | null;
-  }>
-) => {
-  const count = rows.length;
-  // Prisma's _avg ignores NULLs: average only the non-null confidences (and
-  // return null when there are none). Dividing by all rows would diverge from
-  // the real nullable-field aggregate and could hide bugs.
-  const confidences = rows
-    .map(r => r.confidence)
-    .filter((c): c is number => c != null);
-  const avgConfidence = confidences.length
-    ? confidences.reduce((s, c) => s + c, 0) / confidences.length
-    : null;
-  prisma.segmentation.aggregate.mockResolvedValue({
-    _avg: { confidence: avgConfidence },
-    _count: { _all: count },
-  });
-
-  const modelCounts = new Map<string, number>();
-  for (const r of rows) {
-    if (r.model) {
-      modelCounts.set(r.model, (modelCounts.get(r.model) ?? 0) + 1);
-    }
-  }
-  prisma.segmentation.groupBy.mockResolvedValue(
-    [...modelCounts].map(([model, c]) => ({ model, _count: { _all: c } }))
-  );
-
-  const totalPolygons = rows.reduce((s, r) => {
-    try {
-      const parsed = JSON.parse(r.polygons ?? '[]');
-      return s + (Array.isArray(parsed) ? parsed.length : 0);
-    } catch {
-      return s;
-    }
-  }, 0);
-  prisma.$queryRaw.mockResolvedValue([{ total: BigInt(totalPolygons) }]);
-};
-
-// ─── test suite ──────────────────────────────────────────────────────────────
-
-describe('SegmentationService — uncovered paths', () => {
-  let service: SegmentationService;
-  let prismaMock: any;
-  let imageServiceMock: any;
-
-  beforeEach(() => {
-    prismaMock = {
-      segmentation: {
-        findMany: vi.fn(),
-        findUnique: vi.fn(),
-        upsert: vi.fn(),
-        create: vi.fn(),
-        update: vi.fn(),
-        delete: vi.fn(),
-        deleteMany: vi.fn(),
-        aggregate: vi.fn(),
-        groupBy: vi.fn(),
-      },
-      image: {
-        findMany: vi.fn(),
-        findUnique: vi.fn(),
-        count: vi.fn(),
-      },
-      project: {
-        findFirst: vi.fn(),
-        findUnique: vi.fn(),
-      },
-      $transaction: vi.fn(),
-      $queryRaw: vi.fn(),
-    };
-
-    imageServiceMock = {
-      getImageById: vi.fn(),
-      updateSegmentationStatus: vi.fn().mockResolvedValue(undefined),
-    };
-
-    service = new SegmentationService(
-      prismaMock as PrismaClient,
-      imageServiceMock as ImageService
-    );
-
-    vi.clearAllMocks();
-
-    // Re-wire mocks cleared by clearAllMocks()
-    imageServiceMock.updateSegmentationStatus.mockResolvedValue(undefined);
-  });
-
-  // ─── saveSegmentationResults (public wrapper) ───────────────────────────────
-
-  describe('saveSegmentationResults', () => {
-    it('upserts with polygon count and average confidence', async () => {
-      const segId = 'seg-uuid';
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: segId });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [makePolygon() as any],
-        'hrnet',
-        0.5,
-        null,
-        2000, // processingTime ms → stored as seconds / 1000
-        800,
-        600,
-        'user-1'
-      );
-
-      expect(prismaMock.segmentation.upsert).toHaveBeenCalledOnce();
-      const call = prismaMock.segmentation.upsert.mock.calls[0][0];
-      // processingTime: 2000 ms / 1000 = 2 s × 1000 back to ms = 2000
-      expect(call.update.imageWidth).toBe(800);
-      expect(call.update.imageHeight).toBe(600);
-      expect(call.create.model).toBe('hrnet');
-    });
-
-    it('filters out polygons with fewer than 3 points', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-2' });
-
-      const badPolygon = makePolygon({
-        points: [
-          { x: 0, y: 0 },
-          { x: 1, y: 1 },
-        ], // only 2 points
-      });
-      const goodPolygon = makePolygon({ id: 'poly-good' });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [badPolygon as any, goodPolygon as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const call = prismaMock.segmentation.upsert.mock.calls[0][0];
-      const saved: unknown[] = JSON.parse(call.update.polygons);
-      // Only the good polygon should have been kept
-      expect(saved).toHaveLength(1);
-      expect((saved[0] as any).id).toBe('poly-good');
-    });
-
-    it('filters out polygons with invalid point coordinates (NaN)', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-3' });
-
-      const nanPolygon = makePolygon({
-        points: [
-          { x: NaN, y: 0 },
-          { x: 1, y: 1 },
-          { x: 2, y: 2 },
-        ],
-      });
-      const goodPolygon = makePolygon({ id: 'poly-ok' });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [nanPolygon as any, goodPolygon as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const call = prismaMock.segmentation.upsert.mock.calls[0][0];
-      const saved: unknown[] = JSON.parse(call.update.polygons);
-      expect(saved).toHaveLength(1);
-      expect((saved[0] as any).id).toBe('poly-ok');
-    });
-
-    it('filters out polygons with invalid type', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-4' });
-
-      const badType = makePolygon({ type: 'unknown' });
-      const good = makePolygon({ id: 'good' });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [badType as any, good as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const saved: unknown[] = JSON.parse(
-        prismaMock.segmentation.upsert.mock.calls[0][0].update.polygons
-      );
-      expect(saved).toHaveLength(1);
-      expect((saved[0] as any).id).toBe('good');
-    });
-
-    it('clears invalid parent_id references', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-5' });
-
-      // Internal polygon referencing a parent that is not in the list
-      const internal = makePolygon({
-        id: 'internal-1',
-        type: 'internal',
-        parentIds: ['does-not-exist'],
-      });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [internal as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const saved: unknown[] = JSON.parse(
-        prismaMock.segmentation.upsert.mock.calls[0][0].update.polygons
-      );
-      // parent_id should have been cleared (dangling reference)
-      expect((saved[0] as any).parent_id).toBeUndefined();
-    });
-
-    it('merges polylines (sperm model) into polygon list before saving', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-6' });
-
-      const polygon = makePolygon({ id: 'closed-1' }) as any;
-      const polyline = {
-        id: 'open-1',
-        points: [
-          { x: 0, y: 0 },
-          { x: 5, y: 5 },
-        ],
-        type: 'external',
-        area: 0,
-        confidence: 0.8,
-        geometry: 'polyline',
-        partClass: 'head',
-        instanceId: 'sperm-1',
-      };
-
-      // The public wrapper wraps everything into a SegmentationResponse where
-      // polylines come from `segmentationResult.polylines`. We test this via
-      // the private internal path by passing polylines via the response object.
-      // Since the public method doesn't expose polylines directly, we call
-      // saveSegmentationResults with only polygons and verify the merge happens
-      // via the internal code path when called from requestSegmentation.
-      // Instead we use the upsert shape to verify geometry is preserved.
-      await service.saveSegmentationResults(
-        'img-1',
-        [polygon, polyline as any],
-        'sperm',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const saved: unknown[] = JSON.parse(
-        prismaMock.segmentation.upsert.mock.calls[0][0].update.polygons
-      );
-      const savedPolyline = saved.find((p: any) => p.id === 'open-1') as any;
-      expect(savedPolyline).toBeDefined();
-      expect(savedPolyline.geometry).toBe('polyline');
-      expect(savedPolyline.partClass).toBe('head');
-      expect(savedPolyline.instanceId).toBe('sperm-1');
-    });
-
-    it('auto-assigns id to polygons that arrive without one', async () => {
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-7' });
-
-      const noId = {
-        points: [
-          { x: 0, y: 0 },
-          { x: 5, y: 0 },
-          { x: 5, y: 5 },
-        ],
-        type: 'external',
-        area: 12,
-        confidence: 0.7,
-      };
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [noId as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null,
-        null,
-        'user-1'
-      );
-
-      const saved: unknown[] = JSON.parse(
-        prismaMock.segmentation.upsert.mock.calls[0][0].update.polygons
-      );
-      expect((saved[0] as any).id).toBe('polygon_1');
-    });
-
-    it('throws when image_size is missing from the internal response shape', async () => {
-      // This tests the validation branch in saveSegmentationResultsInternal
-      // via the public method which always provides image_size — so we
-      // need to mock the upsert to not be called.
-      // The public wrapper always sets image_size, but let's confirm zero
-      // dimensions are stored when width/height are null.
-      prismaMock.segmentation.upsert.mockResolvedValue({ id: 'seg-8' });
-
-      await service.saveSegmentationResults(
-        'img-1',
-        [makePolygon() as any],
-        'hrnet',
-        0.5,
-        null,
-        null,
-        null, // null width
-        null, // null height
-        'user-1'
-      );
-
-      const call = prismaMock.segmentation.upsert.mock.calls[0][0];
-      expect(call.update.imageWidth).toBe(0);
-      expect(call.update.imageHeight).toBe(0);
-    });
-  });
-
-  // ─── getSegmentationResults ─────────────────────────────────────────────────
-
-  describe('getSegmentationResults', () => {
-    const imageId = 'img-get-1';
-    const userId = 'user-get-1';
-
-    const makeSegRow = (
-      polygons: unknown[],
-      extra: Record<string, unknown> = {}
-    ) => ({
-      id: 'seg-row-1',
-      imageId,
-      polygons: JSON.stringify(polygons),
-      model: 'hrnet',
-      threshold: 0.5,
-      confidence: 0.85,
-      processingTime: 3000,
-      imageWidth: 1024,
-      imageHeight: 768,
-      updatedAt: new Date('2026-01-01T12:00:00Z'),
-      ...extra,
-    });
-
-    beforeEach(() => {
-      imageServiceMock.getImageById.mockResolvedValue({
-        id: imageId,
-        name: 'test.png',
-      });
-    });
-
-    it('returns null when imageService returns null (no access)', async () => {
-      imageServiceMock.getImageById.mockResolvedValue(null);
-      const result = await service.getSegmentationResults(imageId, userId);
-      expect(result).toBeNull();
-    });
-
-    it('returns null when no segmentation row exists', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(null);
-      const result = await service.getSegmentationResults(imageId, userId);
-      expect(result).toBeNull();
-    });
-
-    it('preserves trackId on polylines (MT cross-frame identity)', async () => {
-      const polyWithTrack = {
-        id: 'mt-1',
-        points: [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        type: 'external',
-        area: 0,
-        confidence: 0.9,
-        geometry: 'polyline',
-        trackId: 'track-abc',
-        name: 'MT-1',
-      };
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([polyWithTrack])
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      const poly = result!.polygons[0];
-      expect(poly.trackId).toBe('track-abc');
-      expect(poly.name).toBe('MT-1');
-    });
-
-    it('strips _embedding from response (server-only blob)', async () => {
-      const polyWithEmbedding = {
-        id: 'mt-2',
-        points: [
-          { x: 0, y: 0 },
-          { x: 1, y: 1 },
-        ],
-        type: 'external',
-        area: 0,
-        confidence: 0.9,
-        geometry: 'polyline',
-        trackId: 'track-xyz',
-        _embedding: Array(32).fill(0.1), // should NOT appear in response
-      };
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([polyWithEmbedding])
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      const poly = result!.polygons[0] as any;
-      expect(poly._embedding).toBeUndefined();
-      expect(poly.trackId).toBe('track-xyz');
-    });
-
-    it('converts parent_id to parentIds array', async () => {
-      const internal = {
-        id: 'int-1',
-        points: [
-          { x: 0, y: 0 },
-          { x: 2, y: 0 },
-          { x: 2, y: 2 },
-        ],
-        type: 'internal',
-        area: 4,
-        confidence: 0.8,
-        parent_id: 'ext-parent-1',
-      };
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([internal])
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      expect(result!.polygons[0].parentIds).toEqual(['ext-parent-1']);
-    });
-
-    it('surfaces updatedAt as ISO string for resegment poll', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([makePolygon()])
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      expect(result!.updatedAt).toBe('2026-01-01T12:00:00.000Z');
-    });
-
-    it('surfaces imageWidth / imageHeight from the DB row', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([makePolygon()])
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      expect(result!.imageWidth).toBe(1024);
-      expect(result!.imageHeight).toBe(768);
-      expect(result!.image_size).toEqual({ width: 1024, height: 768 });
-    });
-
-    it('converts processingTime from ms (DB) to seconds (response)', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(
-        makeSegRow([makePolygon()], { processingTime: 4500 })
-      );
-
-      const result = await service.getSegmentationResults(imageId, userId);
-
-      expect(result!.processing_time).toBe(4.5);
-    });
-  });
-
-  // ─── updateSegmentationResults ──────────────────────────────────────────────
-
-  describe('updateSegmentationResults', () => {
-    const imageId = 'img-upd-1';
-    const userId = 'user-upd-1';
-
-    const makePolygons = () => [
-      {
-        id: 'p-1',
-        points: [
-          { x: 0, y: 0 },
-          { x: 5, y: 0 },
-          { x: 5, y: 5 },
-        ],
-        type: 'external' as const,
-        area: 12,
-        confidence: 0.88,
-        parentIds: undefined,
-        geometry: undefined,
-        partClass: undefined,
-        instanceId: undefined,
-        trackId: undefined,
-        name: undefined,
-      },
-    ];
-
-    beforeEach(() => {
-      imageServiceMock.getImageById.mockResolvedValue({
-        id: imageId,
-        name: 'test.png',
-        projectId: 'proj-1',
-        parentVideoId: null, // standalone image — no cross-frame propagation
-      });
-    });
-
-    it('throws when image not found (no access)', async () => {
-      imageServiceMock.getImageById.mockResolvedValue(null);
-
-      await expect(
-        service.updateSegmentationResults(imageId, makePolygons(), userId)
-      ).rejects.toThrow('Image not found or no access');
-    });
-
-    it('creates a new segmentation row when none exists', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(null);
-      prismaMock.segmentation.create.mockResolvedValue({
-        id: 'new-seg-1',
-        imageId,
-        model: 'manual',
-        threshold: 0.5,
-        confidence: 0.88,
-        imageWidth: null,
-        imageHeight: null,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      });
-
-      const result = await service.updateSegmentationResults(
-        imageId,
-        makePolygons(),
-        userId
-      );
-
-      expect(prismaMock.segmentation.create).toHaveBeenCalledOnce();
-      expect(prismaMock.segmentation.create.mock.calls[0][0].data.model).toBe(
-        'manual'
-      );
-      expect(result.status).toBe('completed');
-    });
-
-    it('creates segmentation with dimensions when provided', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(null);
-      prismaMock.segmentation.create.mockResolvedValue({
-        id: 'seg-dim',
-        imageId,
-        model: 'manual',
-        threshold: 0.5,
-        confidence: 0.88,
-        imageWidth: 1024,
-        imageHeight: 768,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      });
-
-      const result = await service.updateSegmentationResults(
-        imageId,
-        makePolygons(),
-        userId,
-        1024,
-        768
-      );
-
-      const createCall = prismaMock.segmentation.create.mock.calls[0][0].data;
-      expect(createCall.imageWidth).toBe(1024);
-      expect(createCall.imageHeight).toBe(768);
-      expect(result.imageWidth).toBe(1024);
-    });
-
-    it('updates existing segmentation and runs transaction', async () => {
-      const existingSeg = {
-        id: 'existing-seg-1',
-        imageId,
-        polygons: JSON.stringify([makePolygon()]),
-        model: 'hrnet',
-        threshold: 0.5,
-        confidence: 0.9,
-        imageWidth: 800,
-        imageHeight: 600,
-      };
-      prismaMock.segmentation.findUnique.mockResolvedValue(existingSeg);
-      // No sibling frames (standalone image)
-      prismaMock.image.findMany.mockResolvedValue([]);
-
-      const updatedRow = {
-        id: 'existing-seg-1',
-        imageId,
-        model: 'hrnet',
-        threshold: 0.5,
-        confidence: 0.88,
-        imageWidth: 800,
-        imageHeight: 600,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      };
-      prismaMock.$transaction.mockResolvedValue([updatedRow]);
-
-      const result = await service.updateSegmentationResults(
-        imageId,
-        makePolygons(),
-        userId
-      );
-
-      expect(prismaMock.$transaction).toHaveBeenCalledOnce();
-      expect(result.id).toBe('existing-seg-1');
-      expect(result.status).toBe('completed');
-    });
-
-    it('converts parentIds array to parent_id in DB polygon', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(null);
-      prismaMock.segmentation.create.mockResolvedValue({
-        id: 'seg-parent',
-        imageId,
-        model: 'manual',
-        threshold: 0.5,
-        confidence: 0.8,
-        imageWidth: null,
-        imageHeight: null,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      });
-
-      const polyWithParent = {
-        id: 'int-p',
-        points: [
-          { x: 0, y: 0 },
-          { x: 1, y: 0 },
-          { x: 1, y: 1 },
-        ],
-        type: 'internal' as const,
-        area: 1,
-        confidence: 0.75,
-        parentIds: ['ext-id-1'],
-      };
-
-      await service.updateSegmentationResults(
-        imageId,
-        [polyWithParent] as any,
-        userId
-      );
-
-      const createCall = prismaMock.segmentation.create.mock.calls[0][0].data;
-      const dbPolygons = JSON.parse(createCall.polygons);
-      expect(dbPolygons[0].parent_id).toBe('ext-id-1');
-      // Original parentIds key should not be in the DB shape
-      expect(dbPolygons[0].parentIds).toBeUndefined();
-    });
-
-    it('preserves trackId and name on update (MT cross-frame identity)', async () => {
-      prismaMock.segmentation.findUnique.mockResolvedValue(null);
-      prismaMock.segmentation.create.mockResolvedValue({
-        id: 'seg-mt',
-        imageId,
-        model: 'manual',
-        threshold: 0.5,
-        confidence: 0.9,
-        imageWidth: null,
-        imageHeight: null,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      });
-
-      const mtPolyline = {
-        id: 'mt-poly',
-        points: [
-          { x: 0, y: 0 },
-          { x: 10, y: 10 },
-        ],
-        type: 'external' as const,
-        area: 0,
-        confidence: 0.9,
-        geometry: 'polyline' as const,
-        trackId: 'track-42',
-        name: 'MT-42',
-      };
-
-      await service.updateSegmentationResults(
-        imageId,
-        [mtPolyline] as any,
-        userId
-      );
-
-      const dbPolygons = JSON.parse(
-        prismaMock.segmentation.create.mock.calls[0][0].data.polygons
-      );
-      expect(dbPolygons[0].trackId).toBe('track-42');
-      expect(dbPolygons[0].name).toBe('MT-42');
-    });
-  });
-
-  // ─── deleteSegmentationResults ──────────────────────────────────────────────
-
-  describe('deleteSegmentationResults', () => {
-    it('deletes segmentation and resets image status', async () => {
-      imageServiceMock.getImageById.mockResolvedValue({
-        id: 'img-del',
-        name: 'test.png',
-      });
-      prismaMock.segmentation.deleteMany.mockResolvedValue({ count: 1 });
-
-      await service.deleteSegmentationResults('img-del', 'user-1');
-
-      expect(prismaMock.segmentation.deleteMany).toHaveBeenCalledWith({
-        where: { imageId: 'img-del' },
-      });
-      expect(imageServiceMock.updateSegmentationStatus).toHaveBeenCalledWith(
-        'img-del',
-        'no_segmentation',
-        'user-1'
-      );
-    });
-
-    it('throws when image not found or no access', async () => {
-      imageServiceMock.getImageById.mockResolvedValue(null);
-
-      await expect(
-        service.deleteSegmentationResults('bad-img', 'user-1')
-      ).rejects.toThrow('Image not found or no access');
-
-      expect(prismaMock.segmentation.deleteMany).not.toHaveBeenCalled();
-    });
-  });
-
-  // ─── getProjectSegmentationStats ─────────────────────────────────────────────
-
-  describe('getProjectSegmentationStats', () => {
-    it('throws when project not found', async () => {
-      prismaMock.project.findFirst.mockResolvedValue(null);
-
-      await expect(
-        service.getProjectSegmentationStats('bad-proj', 'user-1')
-      ).rejects.toThrow('Project not found or no access');
-    });
-
-    it('tallies model usage counts correctly', async () => {
-      prismaMock.project.findFirst.mockResolvedValue({ id: 'proj-1' });
-      prismaMock.image.count.mockResolvedValue(5);
-      mockSegStatsFromRows(prismaMock, [
-        {
-          polygons: JSON.stringify([makePolygon(), makePolygon()]),
-          confidence: 0.9,
-          model: 'hrnet',
-        },
-        {
-          polygons: JSON.stringify([makePolygon()]),
-          confidence: 0.8,
-          model: 'hrnet',
-        },
-        {
-          polygons: JSON.stringify([makePolygon()]),
-          confidence: 0.7,
-          model: 'cbam_resunet',
-        },
-      ]);
-
-      const stats = await service.getProjectSegmentationStats(
-        'proj-1',
-        'user-1'
-      );
-
-      expect(stats.totalImages).toBe(5);
-      expect(stats.processedImages).toBe(3);
-      expect(stats.totalPolygons).toBe(4);
-      expect(stats.models.hrnet).toBe(2);
-      expect(stats.models.cbam_resunet).toBe(1);
-      expect(stats.averageConfidence).toBeCloseTo((0.9 + 0.8 + 0.7) / 3, 5);
-    });
-
-    it('returns zeros when no segmentation data exists', async () => {
-      prismaMock.project.findFirst.mockResolvedValue({ id: 'proj-empty' });
-      prismaMock.image.count.mockResolvedValue(3);
-      mockSegStatsFromRows(prismaMock, []);
-
-      const stats = await service.getProjectSegmentationStats(
-        'proj-empty',
-        'user-1'
-      );
-
-      expect(stats.processedImages).toBe(0);
-      expect(stats.totalPolygons).toBe(0);
-      expect(stats.averageConfidence).toBe(0);
-      expect(stats.models).toEqual({});
-    });
-  });
-});
