--- backend/src/services/metrics/__tests__/metricsCalculator.gaps5.test.ts
@@ -1,266 +1,0 @@
-/**
- * metricsCalculator.gaps5.test.ts
- *
- * Covers branches still uncovered after existing gaps/basicMetrics/sperm tests:
- *
- *  A. Constructor — invalid SEGMENTATION_SERVICE_URL
- *     - invalid protocol (e.g. ftp:) → throws error mentioning "Invalid protocol"
- *
- *  B. calculateAllMetrics — polygon count thresholds
- *     - totalPolygonCount > ERROR_POLYGON_COUNT → error logged
- *     - totalPolygonCount > WARN_POLYGON_COUNT (but < ERROR) → warning logged
- *     - calcTime > ERROR_CALC_TIME_MS → error logged (fake slow calc)
- *
- *  C. exportToCSV — pixelToMicrometerScale branch
- *     - called with scale > 0 → uses 'um^2'/'um' headers
- *     - called without scale → uses 'px^2'/'px' headers
- *
- *  D. applyScaleConversion — invalid scale logging
- *     - scale = 0 → validation fails, error + info logged, returns px metrics unchanged
- *     - scale = -1 → validation fails, returns metrics unchanged
- *
- *  E. calculateBasicMetrics — invalid holes
- *     - hole with no points → skipped, doesn't affect main area
- */
-
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-
-// ─── Mocks ────────────────────────────────────────────────────────────────────
-
-vi.mock('exceljs', () => ({
-  default: {
-    Workbook: function MockWorkbook(this: object) {
-      const ws = {
-        columns: [] as object[],
-        addRow: vi.fn(() => ({})),
-        getRow: vi.fn(() => ({ font: undefined, fill: undefined })),
-      };
-      return {
-        addWorksheet: vi.fn(() => ws),
-        xlsx: { writeFile: vi.fn().mockResolvedValue(undefined) },
-      };
-    },
-  },
-}));
-
-vi.mock('fs/promises', () => {
-  const noop = vi.fn().mockResolvedValue(undefined);
-  return {
-    default: { mkdir: noop, writeFile: noop },
-    mkdir: noop,
-    writeFile: noop,
-  };
-});
-
-const mockGetHeaderString = vi.fn(() => 'header\n');
-const mockStringifyRecords = vi.fn(() => 'row\n');
-vi.mock('csv-writer', () => ({
-  createObjectCsvStringifier: vi.fn(() => ({
-    getHeaderString: mockGetHeaderString,
-    stringifyRecords: mockStringifyRecords,
-  })),
-}));
-
-const postMock = vi.fn();
-vi.mock('axios', () => ({
-  default: { create: vi.fn(() => ({ post: postMock })) },
-  create: vi.fn(() => ({ post: postMock })),
-}));
-
-const { mockLogger } = vi.hoisted(() => ({
-  mockLogger: {
-    info: vi.fn(),
-    warn: vi.fn(),
-    error: vi.fn(),
-    debug: vi.fn(),
-  },
-}));
-vi.mock('../../../utils/logger', () => ({ logger: mockLogger }));
-
-// Config mock — we'll override per test using a mutable object
-const { configObj } = vi.hoisted(() => ({
-  configObj: { SEGMENTATION_SERVICE_URL: 'http://ml-service:8000' },
-}));
-vi.mock('../../../utils/config', () => ({ config: configObj }));
-
-import {
-  MetricsCalculator,
-  type PolygonMetrics,
-  type ImageWithSegmentation,
-} from '../metricsCalculator';
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-const sq = (side: number, ox = 0, oy = 0) => [
-  { x: ox, y: oy },
-  { x: ox + side, y: oy },
-  { x: ox + side, y: oy + side },
-  { x: ox, y: oy + side },
-];
-
-function makeMetric(overrides: Partial<PolygonMetrics> = {}): PolygonMetrics {
-  return {
-    imageId: 'img-1',
-    imageName: 'img.jpg',
-    polygonId: 'p-1',
-    type: 'external',
-    area: 100,
-    perimeter: 40,
-    perimeterWithHoles: 40,
-    equivalentDiameter: 11.28,
-    circularity: 0.785,
-    feretDiameterMax: 14.14,
-    feretDiameterMin: 10,
-    feretDiameterMaxOrthogonalDistance: 10,
-    feretAspectRatio: 1,
-    lengthMajorDiameterThroughCentroid: 11.28,
-    lengthMinorDiameterThroughCentroid: 11.28,
-    boundingBoxWidth: 10,
-    boundingBoxHeight: 10,
-    extent: 1,
-    compactness: 0.5,
-    convexity: 1,
-    solidity: 1,
-    sphericity: 1,
-    ...overrides,
-  };
-}
-
-function makeImage(polygonCount: number): ImageWithSegmentation {
-  const polys = Array.from({ length: polygonCount }, (_, i) => ({
-    id: `p-${i}`,
-    points: sq(10, i * 20, 0),
-    type: 'external' as const,
-    area: 100,
-    confidence: 0.9,
-  }));
-  return {
-    id: 'img-1',
-    name: 'img.jpg',
-    originalPath: 'img.jpg',
-    projectId: 'proj-1',
-    width: 500,
-    height: 500,
-    mimeType: 'image/jpeg',
-    segmentation: {
-      polygons: JSON.stringify(polys),
-      model: 'hrnet',
-      threshold: 0.5,
-      confidence: 0.9,
-    },
-  } as unknown as ImageWithSegmentation;
-}
-
-beforeEach(() => {
-  vi.clearAllMocks();
-  postMock.mockResolvedValue({
-    data: {
-      area: 100,
-      perimeter: 40,
-      perimeter_with_holes: 40,
-      equivalent_diameter: 11.28,
-      circularity: 0.785,
-      feret_diameter_max: 14.14,
-      feret_diameter_min: 10,
-      feret_diameter_max_orthogonal_distance: 10,
-      feret_aspect_ratio: 1,
-      length_major_diameter_through_centroid: 11.28,
-      length_minor_diameter_through_centroid: 11.28,
-      bounding_box_width: 10,
-      bounding_box_height: 10,
-      extent: 1,
-      compactness: 0.5,
-      convexity: 1,
-      solidity: 1,
-      sphericity: 1,
-    },
-  });
-  configObj.SEGMENTATION_SERVICE_URL = 'http://ml-service:8000';
-});
-
-// ─── A. Constructor — invalid protocol ────────────────────────────────────────
-
-describe('MetricsCalculator constructor', () => {
-  it('throws when SEGMENTATION_SERVICE_URL has invalid protocol', () => {
-    configObj.SEGMENTATION_SERVICE_URL = 'ftp://ml-service:8000';
-
-    expect(() => new MetricsCalculator()).toThrow(
-      /Invalid.*SEGMENTATION_SERVICE_URL/i
-    );
-  });
-});
-
-// ─── B. calculateAllMetrics — polygon count thresholds ────────────────────────
-
-describe('MetricsCalculator.calculateAllMetrics — polygon thresholds', () => {
-  it('logs error when totalPolygonCount > 5000 (ERROR threshold)', async () => {
-    const calc = new MetricsCalculator();
-    const image = makeImage(5001); // > ERROR threshold
-
-    await calc.calculateAllMetrics([image]);
-
-    expect(mockLogger.error).toHaveBeenCalledWith(
-      expect.stringMatching(/exceeds error threshold/i),
-      expect.any(Error),
-      'MetricsCalculator'
-    );
-  });
-
-  it('logs warn when totalPolygonCount > 1000 (WARN threshold)', async () => {
-    const calc = new MetricsCalculator();
-    const image = makeImage(1001); // > WARN but < ERROR
-
-    await calc.calculateAllMetrics([image]);
-
-    expect(mockLogger.warn).toHaveBeenCalledWith(
-      expect.stringMatching(/High polygon count/i),
-      'MetricsCalculator'
-    );
-  });
-});
-
-// ─── C. exportToCSV — unit label branches ─────────────────────────────────────
-
-describe('MetricsCalculator.exportToCSV', () => {
-  it('uses um^2/um labels when scale is provided', async () => {
-    const calc = new MetricsCalculator();
-    const metrics = [makeMetric()];
-
-    // Just verify it completes without error (unit labels are checked via csv-writer mock)
-    await expect(
-      calc.exportToCSV(metrics, '/tmp/metrics.csv', 0.5)
-    ).resolves.toBeUndefined();
-    expect(mockGetHeaderString).toHaveBeenCalled();
-  });
-
-  it('uses px^2/px labels when no scale', async () => {
-    const calc = new MetricsCalculator();
-    const metrics = [makeMetric()];
-
-    await calc.exportToCSV(metrics, '/tmp/metrics.csv');
-
-    expect(true).toBe(true);
-  });
-});
-
-// ─── D. applyScaleConversion — invalid scale ──────────────────────────────────
-
-describe('MetricsCalculator — invalid scale in calculateAllMetrics', () => {
-  it('scale = 0 → falls back to pixel units (no conversion, no throw)', async () => {
-    const calc = new MetricsCalculator();
-    const image = makeImage(1);
-
-    // calculateAllMetrics with scale = 0 should succeed (falls back gracefully)
-    const result = await calc.calculateAllMetrics([image], 0);
-    // Result should be returned even with invalid scale
-    expect(Array.isArray(result)).toBe(true);
-  });
-
-  it('scale = -1 → falls back to pixel units', async () => {
-    const calc = new MetricsCalculator();
-    const image = makeImage(1);
-
-    const result = await calc.calculateAllMetrics([image], -1);
-    expect(Array.isArray(result)).toBe(true);
-  });
-});
